rename utils to pkg
