+++ internal/pkg/auth/errors.go
